python-backend/requirements.txt
fastapifastapi==0.111.0
uvicorn==0.30.1

motor==3.4.0
pymongo==4.7.2
dnspython==2.6.1

pydantic==2.7.4
pydantic-core==2.18.4
typing-extensions==4.12.2

python-dotenv==1.0.1


